model_training/requirements.txt
# COD-Odisha Model Training Requirements

# Core training dependencies
torch==2.0.1
torchvision==0.15.2
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0
scipy>=1.10.0

# Dataset utilities
opencv-python>=4.8.0
imageio>=2.32.0
scikit-image>=0.21.0

# Evaluation metrics
tensorboard>=2.14.0
wandb>=0.15.0

# Optional: For advanced augmentation
albumentations>=1.3.0
